Move WindowMessage to its own file
