+++ src/tests/unit/tests/injected/frameCommunicators/window-message-handler.test.ts
@@ -4,8 +4,9 @@ import { IMock, It, Mock, MockBehavior, Times } from 'typemoq';
 
 import { WindowUtils } from '../../../../../common/window-utils';
 import { ErrorMessageContent } from '../../../../../injected/frameCommunicators/error-message-content';
+import { WindowMessage } from '../../../../../injected/frameCommunicators/window-message';
 import { FrameMessageResponseCallback, WindowMessageHandler } from '../../../../../injected/frameCommunicators/window-message-handler';
-import { WindowMessage, WindowMessageMarshaller } from '../../../../../injected/frameCommunicators/window-message-marshaller';
+import { WindowMessageMarshaller } from '../../../../../injected/frameCommunicators/window-message-marshaller';
 
 describe('WindowMessageHandlerTests', () => {
     let testSubject: WindowMessageHandler;
